+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## $v
+## v160731.37352
 
 - Adding `Classes\SCore\WidgetForm{}`
 - Adding `Classes\SCore\Base\Widget{}`